+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-07-21 11:29:21
-// Commit: test(state): implement widget integration tests for state
+// Updated at: 2026-07-21 15:53:37
+// Commit: style(state): harmonize light and dark theme palette for state
 
 class stateComponentState {
-  final String commitRef = '2985';
-  final String lastMessage = 'test(state): implement widget integration tests for state';
-  final DateTime timestamp = DateTime.parse('2026-07-21T11:29:21+05:00');
+  final String commitRef = '3002';
+  final String lastMessage = 'style(state): harmonize light and dark theme palette for state';
+  final DateTime timestamp = DateTime.parse('2026-07-21T15:53:37+05:00');
 }